+++ project/Dependencies.scala
@@ -5,7 +5,7 @@ import Keys._
 object Dependencies {
   object versions {
     val grpc                 = "1.62.2"
-    val protobuf             = "3.19.6"
+    val protobuf             = "4.26.0"
     val silencer             = "1.7.16"
     val collectionCompat     = "2.11.0"
     val coursier             = "2.1.9"